Addressing review comments: Updated tests to add more asserts and moved ITs to IcapClientIT
